+++ src/status.rs
@@ -49,7 +49,8 @@ pub struct State {
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
 pub struct Event {
     pub name: String,
-    pub _type: String,
+    #[serde(rename = "type")]
+    pub type_: String,
     pub timestamp: u64,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub extra: Option<String>,
@@ -107,8 +108,8 @@ pub struct Contact {
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
 pub struct Feed {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub _type: Option<String>,
+    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
+    pub type_: Option<String>,
     pub url: String,
 }
 
@@ -133,7 +134,8 @@ pub struct Cache {
 pub struct RadioShow {
     pub name: String,
     pub url: String,
-    pub _type: String,
+    #[serde(rename = "type")]
+    pub type_: String,
     pub start: String,
     pub end: String,
 }
@@ -618,15 +620,15 @@ mod test {
     #[test]
     fn serialize_skip_none() {
         let f1 = Feed {
-            _type: Some("rss".to_string()),
+            type_: Some("rss".to_string()),
             url: "https://some/rss.xml".to_string(),
         };
         let f2 = Feed {
-            _type: None,
+            type_: None,
             url: "https://some/rss.xml".to_string(),
         };
         assert_eq!(to_string(&f1).unwrap(),
-            "{\"_type\":\"rss\",\"url\":\"https://some/rss.xml\"}".to_string());
+            "{\"type\":\"rss\",\"url\":\"https://some/rss.xml\"}".to_string());
         assert_eq!(to_string(&f2).unwrap(),
             "{\"url\":\"https://some/rss.xml\"}".to_string());
     }